Refactor run.sh script in a backward-compatible manner
First of all, I would leave all the building/cleaning to Make,
because there is no need to reinvent building mechanics.
But since run.sh is already used for building/cleaning/testing, I
decided not to disrespect things as they are and just refactor
the code, allowing usage in the original way.

What are the enhancements?
1. Extra caution is used to make sure we do not do some unintended stuff:
   current directory is compared against script source directory; EUID
   is also checked to make sure we are not root. Usage of -f option
   makes following these conditional optional.
2. Verbose option is added, but for now it is pretty dumb and just prints
   full files' paths. I believe however this option will prove itself
   useful once some custom debug-printing lines are added.
3. Code redundancy is minimized: main part of run.sh is put in the
   shared directory.
4. Code is refactored, (hopefully) improving general readability. Since
   we do not use Make, let's at least have our own targets with
   blackjack and...